refactor(scheduler): extract activation memory and monitor logic
- Extract activation memory management logic from `BaseScheduler` to new `ActivationMemoryManager` class
- Move background task schedule monitoring loop from `BaseScheduler` to `TaskScheduleMonitor`
- Update `BaseScheduler` to delegate activation memory operations to `ActivationMemoryManager`
- Update `BaseScheduler` to use `TaskScheduleMonitor` for metrics monitoring
- Update tests to reflect changes in scheduler component naming (retriever -> post_processor/search_service)
