+++ src/memos/mem_scheduler/schemas/general_schemas.py
@@ -17,6 +17,7 @@ DEFAULT_ACT_MEM_DUMP_PATH = f"{BASE_DIR}/outputs/mem_scheduler/mem_cube_schedule
 DEFAULT_THREAD_POOL_MAX_WORKERS = 50
 DEFAULT_CONSUME_INTERVAL_SECONDS = 0.01
 DEFAULT_CONSUME_BATCH = 3
+DEFAULT_MONITOR_INTERVAL_SECONDS = 15
 DEFAULT_DISPATCHER_MONITOR_CHECK_INTERVAL = 300
 DEFAULT_DISPATCHER_MONITOR_MAX_FAILURES = 2
 DEFAULT_STUCK_THREAD_TOLERANCE = 10
